Fix TPC-H Q20.
The computation of the excess boundary did not yield the same result as
the SQL equivalent: For suppliers that do not ship the part in the
interval, the SQL subquery that computes the shipped quantities will
return an empty set and the sum aggregate will turn that into NULL.
Consequentially, the comparison with NULL in the enclosing query will
fail, eliminating the partsupp entry.

In constrast, the DSH combinator 'sum' mimics the regular Haskell
version and returns 0 for an empty input. The predicate will return True
in this case and the patsupp entry will not be eliminated.

The inconsistency can be fixed by checking explicitly whether the
shipped quantities are empty.